Split out SBD implementations into their own cpp files. Add optional dependency on OpenSSL; required for Rockblock SBD implementation
