+++ src/acomms/modemdriver/iridium_shore_driver.cpp
@@ -55,9 +55,7 @@
 #include "goby/acomms/acomms_constants.h"                  // for BITS_...
 #include "goby/acomms/modemdriver/iridium_driver_common.h" // for OnCal...
 #include "goby/acomms/modemdriver/iridium_shore_rudics.h"  // for RUDIC...
-#include "goby/acomms/modemdriver/iridium_shore_sbd.h"     // for SBDMO...
 #include "goby/acomms/modemdriver/rudics_packet.h"         // for Rudic...
-#include "goby/acomms/protobuf/iridium_sbd_directip.pb.h"  // for Direc...
 #include "goby/time/convert.h"                             // for Syste...
 #include "goby/time/system_clock.h"                        // for Syste...
 #include "goby/time/types.h"                               // for SITime
@@ -68,25 +66,14 @@
 #include "goby/util/debug_logger/flex_ostreambuf.h"        // for DEBUG1
 #include "goby/util/debug_logger/logger_manipulators.h"    // for opera...
 
-#include "goby/util/thirdparty/jwt-cpp/traits/nlohmann-json/defaults.h"
-
-#include "goby/util/thirdparty/jwt-cpp/jwt.h"
-
 #include "iridium_shore_driver.h"
 
 using namespace goby::util::logger;
 using goby::glog;
-using goby::acomms::iridium::protobuf::DirectIPMTHeader;
-using goby::acomms::iridium::protobuf::DirectIPMTPayload;
 
 goby::acomms::IridiumShoreDriver::IridiumShoreDriver() { init_iridium_dccl(); }
 
 goby::acomms::IridiumShoreDriver::~IridiumShoreDriver() = default;
-
-// from https://docs.rock7.com/reference/push-api
-const std::string rockblock_rsa_pubkey = R"(-----BEGIN PUBLIC KEY-----
-MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAlaWAVJfNWC4XfnRx96p9cztBcdQV6l8aKmzAlZdpEcQR6MSPzlgvihaUHNJgKm8t5ShR3jcDXIOI7er30cIN4/9aVFMe0LWZClUGgCSLc3rrMD4FzgOJ4ibD8scVyER/sirRzf5/dswJedEiMte1ElMQy2M6IWBACry9u12kIqG0HrhaQOzc6Tr8pHUWTKft3xwGpxCkV+K1N+9HCKFccbwb8okRP6FFAMm5sBbw4yAu39IVvcSL43Tucaa79FzOmfGs5mMvQfvO1ua7cOLKfAwkhxEjirC0/RYX7Wio5yL6jmykAHJqFG2HT0uyjjrQWMtoGgwv9cIcI7xbsDX6owIDAQAB
------END PUBLIC KEY-----)";
 
 void goby::acomms::IridiumShoreDriver::startup(const protobuf::DriverConfig& cfg)
 {
@@ -105,26 +92,17 @@ void goby::acomms::IridiumShoreDriver::startup(const protobuf::DriverConfig& cfg
     switch (iridium_shore_driver_cfg().sbd_type())
     {
         case goby::acomms::iridium::protobuf::ShoreConfig::SBD_DIRECTIP:
-            directip_mo_sbd_server_.reset(
-                new directip::SBDServer(sbd_io_, iridium_shore_driver_cfg().mo_sbd_server_port()));
+            startup_sbd_directip(cfg);
             break;
 
         case goby::acomms::iridium::protobuf::ShoreConfig::SBD_ROCKBLOCK:
-            if (!iridium_shore_driver_cfg().has_rockblock())
-                glog.is(DIE) &&
-                    glog << group(glog_out_group())
-                         << "Must specify rockblock {} configuration when using SBD_ROCKBLOCK"
-                         << std::endl;
-
-            // use the build-in modem connection for receiving MO messages as HTTP is line-based
-            driver_cfg_.set_connection_type(
-                goby::acomms::protobuf::DriverConfig::CONNECTION_TCP_AS_SERVER);
-
-            if (iridium_shore_driver_cfg().has_mo_sbd_server_port() || !driver_cfg_.has_tcp_port())
-                driver_cfg_.set_tcp_port(iridium_shore_driver_cfg().mo_sbd_server_port());
-
-            // new line for HTTP; JSON message has no newline so use }
-            driver_cfg_.set_line_delimiter("\n|}");
+#ifdef HAS_OPENSSL
+            startup_sbd_rockblock(cfg);
+#else
+            glog.is_die() &&
+                glog << "Rockblock functionality requires OpenSSL. Recompile with -Denable_openssl"
+                     << std::endl;
+#endif
             break;
     }
 
@@ -407,176 +385,14 @@ void goby::acomms::IridiumShoreDriver::receive_sbd_mo()
             break;
 
         case goby::acomms::iridium::protobuf::ShoreConfig::SBD_ROCKBLOCK:
+#ifdef HAS_OPENSSL
             receive_sbd_mo_rockblock();
+#else
+            glog.is_die() &&
+                glog << "Rockblock functionality requires OpenSSL. Recompile with -Denable_openssl"
+                     << std::endl;
+#endif
             break;
-    }
-}
-
-void goby::acomms::IridiumShoreDriver::receive_sbd_mo_directip()
-{
-    try
-    {
-        sbd_io_.poll();
-    }
-    catch (std::exception& e)
-    {
-        glog.is(DEBUG1) && glog << warn << group(glog_in_group())
-                                << "Could not handle SBD receive: " << e.what() << std::endl;
-    }
-
-    auto it = directip_mo_sbd_server_->connections().begin(),
-         end = directip_mo_sbd_server_->connections().end();
-    while (it != end)
-    {
-        const int timeout = 5;
-        if ((*it)->message().data_ready())
-        {
-            glog.is(DEBUG1) && glog << group(glog_in_group()) << "Rx SBD PreHeader: "
-                                    << (*it)->message().pre_header().DebugString() << std::endl;
-            glog.is(DEBUG1) && glog << group(glog_in_group())
-                                    << "Rx SBD Header: " << (*it)->message().header().DebugString()
-                                    << std::endl;
-            glog.is(DEBUG1) && glog << group(glog_in_group())
-                                    << "Rx SBD Payload: " << (*it)->message().body().DebugString()
-                                    << std::endl;
-
-            receive_sbd_mo_data((*it)->message().body().payload());
-            directip_mo_sbd_server_->connections().erase(it++);
-        }
-        else if ((*it)->connect_time() > 0 &&
-                 (time::SystemClock::now().time_since_epoch() / std::chrono::seconds(1) >
-                  ((*it)->connect_time() + timeout)))
-        {
-            glog.is(DEBUG1) && glog << group(glog_in_group())
-                                    << "Removing SBD connection that has timed out:"
-                                    << (*it)->remote_endpoint_str() << std::endl;
-            directip_mo_sbd_server_->connections().erase(it++);
-        }
-        else
-        {
-            ++it;
-        }
-    }
-}
-
-// e.g.,
-
-// POST / HTTP/1.1
-// User-Agent: Rock7PushApi
-// Content-Type: application/json; charset=utf-8
-// Content-Length: 1125
-// Host: gobysoft.org:8080
-// Connection: Keep-Alive
-// Accept-Encoding: gzip
-
-// {"momsn":66,"data":"546865726520617265203130207479706573206f662070656f706c652077686f20756e6465727374616e642062696e617279","serial":14331,"iridium_latitude":75.0001,"iridium_cep":3.0,"JWT":"eyJhbGciOiJSUzI1NiJ9.eyJpc3MiOiJSb2NrIDciLCJpYXQiOjE2OTM4ODg0NzUsImRhdGEiOiI1NDY4NjU3MjY1MjA2MTcyNjUyMDMxMzAyMDc0Nzk3MDY1NzMyMDZmNjYyMDcwNjU2ZjcwNmM2NTIwNzc2ODZmMjA3NTZlNjQ2NTcyNzM3NDYxNmU2NDIwNjI2OTZlNjE3Mjc5IiwiZGV2aWNlX3R5cGUiOiJST0NLQkxPQ0siLCJpbWVpIjoiMzAwNDM0MDYyMDk5NDMwIiwiaXJpZGl1bV9jZXAiOiIzLjAiLCJpcmlkaXVtX2xhdGl0dWRlIjoiNzUuMDAwMSIsImlyaWRpdW1fbG9uZ2l0dWRlIjoiMTU0LjEyMDkiLCJtb21zbiI6IjY2Iiwic2VyaWFsIjoiMTQzMzEiLCJ0cmFuc21pdF90aW1lIjoiMjMtMDktMDUgMDQ6MzM6MjQifQ.XfBpmA_XmkRoK1PPBjYKm-cQzEXIV-OOjt5ODgF6TX5LMhoDlDkAA7JrFiRnCY8zYMzseLn368oBwNR01yIe8bYVIVM5MM0ykkovRve-ubmwdHYvH0Hgiph-a69yeMwzvGOyfW8WFxTJwg8I_UCCWyFwG_FIHk-iZsiAZ42h7AYDk8zpMStH3-jD5I8ymGQXRe9b-QfLlAOFfXMNBKuFRBsMw1Jgbgv5X5ScpkYrVMozOkjMFaLPSclYeE1jodlv_QUVXR-4g_Xczmdff56HnOfBTVuUwdFkr661r5EoXgX3GB-natzZ5XZ-dy8j84GQYOZi5wBmtFalKdJdmHS1rg","imei":"300434062099430","device_type":"ROCKBLOCK","transmit_time":"23-09-05 04:33:24","iridium_longitude":154.1209}
-void goby::acomms::IridiumShoreDriver::receive_sbd_mo_rockblock()
-{
-    std::string line;
-    const std::string start = "POST / HTTP/1.1";
-    while (modem_read(&line))
-    {
-        if (boost::trim_copy(line) == start)
-        {
-            if (rb_msg_ && rb_msg_->state != RockblockHTTPMessage::MessageState::COMPLETE)
-            {
-                glog.is_warn() &&
-                    glog << group(glog_in_group())
-                         << "Received start of new HTTP message without completing last message"
-                         << std::endl;
-            }
-
-            rb_msg_.reset(new RockblockHTTPMessage);
-        }
-        else if (rb_msg_)
-        {
-            if (rb_msg_->state == RockblockHTTPMessage::MessageState::COMPLETE)
-            {
-                glog.is_warn() && glog << group(glog_in_group())
-                                       << "Received data after complete message, ignoring."
-                                       << std::endl;
-                continue;
-            }
-            else if (rb_msg_->state == RockblockHTTPMessage::MessageState::HEADER)
-            {
-                if (line == "\r\n")
-                {
-                    rb_msg_->state = RockblockHTTPMessage::MessageState::BODY;
-                    for (const auto& h_p : rb_msg_->header)
-                        glog.is_debug2() && glog << group(glog_in_group()) << "Header ["
-                                                 << h_p.first << ":" << h_p.second << "]"
-                                                 << std::endl;
-                }
-                else
-                {
-                    auto colon_pos = line.find(":");
-                    if (colon_pos == std::string::npos)
-                    {
-                        glog.is_warn() && glog << group(glog_in_group())
-                                               << "Received header field without colon, ignoring"
-                                               << std::endl;
-                        continue;
-                    }
-                    std::string key = line.substr(0, colon_pos);
-                    std::string value = line.substr(colon_pos);
-
-                    boost::trim_if(key, boost::is_any_of(": \r\n"));
-                    boost::trim_if(value, boost::is_any_of(": \r\n"));
-                    rb_msg_->header.insert(std::make_pair(key, value));
-                }
-            }
-            else if (rb_msg_->state == RockblockHTTPMessage::MessageState::BODY)
-            {
-                rb_msg_->body += line;
-                if (rb_msg_->header.count("Content-Length"))
-                {
-                    auto length = goby::util::as<int>(rb_msg_->header.at("Content-Length"));
-                    if (rb_msg_->body.size() == length)
-                    {
-                        rb_msg_->state = RockblockHTTPMessage::MessageState::COMPLETE;
-
-                        try
-                        {
-                            auto json_data = nlohmann::json::parse(rb_msg_->body);
-                            glog.is_debug1() && glog << "Received valid JSON message: "
-                                                     << json_data.dump(2) << std::endl;
-
-                            auto verify = jwt::verify()
-                                              .allow_algorithm(jwt::algorithm::rs256(
-                                                  rockblock_rsa_pubkey, "", "", ""))
-                                              .with_issuer("Rock 7");
-                            auto decoded = jwt::decode(json_data["JWT"].get<std::string>());
-                            try
-                            {
-                                verify.verify(decoded);
-                                receive_sbd_mo_data(goby::util::hex_decode(json_data["data"]));
-                            }
-                            catch (const std::exception& e)
-                            {
-                                glog.is_warn() && glog << "Discarding message: could not verify "
-                                                          "Rockblock JWT against public key: "
-                                                       << e.what() << std::endl;
-                            }
-
-                            modem_write("HTTP/1.1 200 OK\r\n");
-                            modem_write("Content-Length: 0\r\n");
-                            modem_write("Connection: close\r\n\r\n");
-                        }
-                        catch (std::exception& e)
-                        {
-                            glog.is_warn() && glog << group(glog_in_group())
-                                                   << "Failed to parse JSON: " << e.what()
-                                                   << ", data: " << rb_msg_->body << std::endl;
-                        }
-                    }
-                }
-                else
-                {
-                    glog.is_warn() && glog << group(glog_in_group())
-                                           << "No Content-Length in header" << std::endl;
-                }
-            }
-        }
     }
 }
 
@@ -605,164 +421,20 @@ void goby::acomms::IridiumShoreDriver::receive_sbd_mo_data(const std::string& da
 void goby::acomms::IridiumShoreDriver::send_sbd_mt(const std::string& bytes,
                                                    const std::string& imei)
 {
-    try
+    switch (iridium_shore_driver_cfg().sbd_type())
     {
-        using boost::asio::ip::tcp;
+        case goby::acomms::iridium::protobuf::ShoreConfig::SBD_DIRECTIP:
+            send_sbd_mt_directip(bytes, imei);
+            break;
 
-        boost::asio::io_service io_service;
-
-        tcp::resolver resolver(io_service);
-        tcp::resolver::query query(
-            iridium_shore_driver_cfg().mt_sbd_server_address(),
-            goby::util::as<std::string>(iridium_shore_driver_cfg().mt_sbd_server_port()),
-            boost::asio::ip::resolver_query_base::numeric_service);
-        tcp::resolver::iterator endpoint_iterator = resolver.resolve(query);
-        tcp::resolver::iterator end;
-
-        tcp::socket socket(io_service);
-        boost::system::error_code error = boost::asio::error::host_not_found;
-        while (error && endpoint_iterator != end)
-        {
-            socket.close();
-            socket.connect(*endpoint_iterator++, error);
-        }
-        if (error)
-            throw boost::system::system_error(error);
-
-        boost::asio::write(socket, boost::asio::buffer(create_sbd_mt_data_message(bytes, imei)));
-
-        directip::SBDMTConfirmationMessageReader message(socket);
-        boost::asio::async_read(
-            socket, boost::asio::buffer(message.data()),
-            boost::asio::transfer_at_least(directip::SBDMessageReader::PRE_HEADER_SIZE),
-            boost::bind(&directip::SBDMessageReader::pre_header_handler, &message,
-                        boost::placeholders::_1, boost::placeholders::_2));
-
-        double start_time = time::SystemClock::now().time_since_epoch() / std::chrono::seconds(1);
-        const int timeout = 5;
-
-        while (!message.data_ready() &&
-               (start_time + timeout >
-                time::SystemClock::now().time_since_epoch() / std::chrono::seconds(1)))
-            io_service.poll();
-
-        if (message.data_ready())
-        {
-            glog.is(DEBUG1) && glog << "Tx SBD Confirmation: " << message.confirm().DebugString()
-                                    << std::endl;
-        }
-        else
-        {
-            glog.is(WARN) && glog << "Timeout waiting for confirmation message from DirectIP server"
-                                  << std::endl;
-        }
+        case goby::acomms::iridium::protobuf::ShoreConfig::SBD_ROCKBLOCK:
+#ifdef HAS_OPENSSL
+            send_sbd_mt_rockblock(bytes, imei);
+#else
+            glog.is_die() &&
+                glog << "Rockblock functionality requires OpenSSL. Recompile with -Denable_openssl"
+                     << std::endl;
+#endif
+            break;
     }
-    catch (std::exception& e)
-    {
-        glog.is(WARN) && glog << "Could not sent MT SBD message: " << e.what() << std::endl;
-    }
-}
-
-std::string goby::acomms::IridiumShoreDriver::create_sbd_mt_data_message(const std::string& bytes,
-                                                                         const std::string& imei)
-{
-    enum
-    {
-        PRE_HEADER_SIZE = 3,
-        BITS_PER_BYTE = 8,
-        IEI_SIZE = 3,
-        HEADER_SIZE = 21
-    };
-
-    enum
-    {
-        IEI_MT_HEADER = 0x41,
-        IEI_MT_PAYLOAD = 0x42
-    };
-
-    static int i = 0;
-    DirectIPMTHeader header;
-    header.set_iei(IEI_MT_HEADER);
-    header.set_length(HEADER_SIZE);
-    header.set_client_id(i++);
-    header.set_imei(imei);
-
-    enum
-    {
-        DISP_FLAG_FLUSH_MT_QUEUE = 0x01,
-        DISP_FLAG_SEND_RING_ALERT_NO_MTM = 0x02,
-        DISP_FLAG_UPDATE_SSD_LOCATION = 0x08,
-        DISP_FLAG_HIGH_PRIORITY_MESSAGE = 0x10,
-        DISP_FLAG_ASSIGN_MTMSN = 0x20
-    };
-
-    header.set_disposition_flags(DISP_FLAG_FLUSH_MT_QUEUE);
-
-    std::string header_bytes(IEI_SIZE + HEADER_SIZE, '\0');
-
-    std::string::size_type pos = 0;
-    enum
-    {
-        HEADER_IEI = 1,
-        HEADER_LENGTH = 2,
-        HEADER_CLIENT_ID = 3,
-        HEADER_IMEI = 4,
-        HEADER_DISPOSITION_FLAGS = 5
-    };
-
-    for (int field = HEADER_IEI; field <= HEADER_DISPOSITION_FLAGS; ++field)
-    {
-        switch (field)
-        {
-            case HEADER_IEI: header_bytes[pos++] = header.iei() & 0xff; break;
-
-            case HEADER_LENGTH:
-                header_bytes[pos++] = (header.length() >> BITS_PER_BYTE) & 0xff;
-                header_bytes[pos++] = (header.length()) & 0xff;
-                break;
-
-            case HEADER_CLIENT_ID:
-                header_bytes[pos++] = (header.client_id() >> 3 * BITS_PER_BYTE) & 0xff;
-                header_bytes[pos++] = (header.client_id() >> 2 * BITS_PER_BYTE) & 0xff;
-                header_bytes[pos++] = (header.client_id() >> BITS_PER_BYTE) & 0xff;
-                header_bytes[pos++] = (header.client_id()) & 0xff;
-                break;
-
-            case HEADER_IMEI:
-                header_bytes.replace(pos, 15, header.imei());
-                pos += 15;
-                break;
-
-            case HEADER_DISPOSITION_FLAGS:
-                header_bytes[pos++] = (header.disposition_flags() >> BITS_PER_BYTE) & 0xff;
-                header_bytes[pos++] = (header.disposition_flags()) & 0xff;
-                break;
-        }
-    }
-
-    DirectIPMTPayload payload;
-    payload.set_iei(IEI_MT_PAYLOAD);
-    payload.set_length(bytes.size());
-    payload.set_payload(bytes);
-
-    std::string payload_bytes(IEI_SIZE + bytes.size(), '\0');
-    payload_bytes[0] = payload.iei();
-    payload_bytes[1] = (payload.length() >> BITS_PER_BYTE) & 0xff;
-    payload_bytes[2] = (payload.length()) & 0xff;
-    payload_bytes.replace(3, payload.payload().size(), payload.payload());
-
-    // Protocol Revision Number (1 byte) == 1
-    // Overall Message Length (2 bytes)
-    int overall_length = header_bytes.size() + payload_bytes.size();
-    std::string pre_header_bytes(PRE_HEADER_SIZE, '\0');
-    pre_header_bytes[0] = 1;
-    pre_header_bytes[1] = (overall_length >> BITS_PER_BYTE) & 0xff;
-    pre_header_bytes[2] = (overall_length)&0xff;
-
-    glog.is(DEBUG1) && glog << "Tx SBD PreHeader: " << goby::util::hex_encode(pre_header_bytes)
-                            << std::endl;
-    glog.is(DEBUG1) && glog << "Tx SBD Header: " << header.DebugString() << std::endl;
-    glog.is(DEBUG1) && glog << "Tx SBD Payload: " << payload.DebugString() << std::endl;
-
-    return pre_header_bytes + header_bytes + payload_bytes;
 }
